coodrinator dashboard with request detail view and approval login functionality
